+++ packages/pattern/mod.ts
@@ -24,6 +24,6 @@
  *
  * @module
  */
-export type { CACell } from "./types.ts";
+export type { CACell, Position } from "./types.ts";
 export { CACellList } from "./ca-cell-list.ts";
 export { BoundingRect } from "./bounding-rect.ts";