+++ packages/pattern/types.ts
@@ -1,18 +1,20 @@
+export type Position = {
+  /**
+   * x coordinate
+   */
+  readonly x: number;
+  /**
+   * y coordinate
+   */
+  readonly y: number;
+};
+
 /**
  * A cell of cellular automaton
  */
 export type CACell = {
   /** Position of the cell */
-  readonly position: {
-    /**
-     * x coordinate
-     */
-    readonly x: number;
-    /**
-     * y coordinate
-     */
-    readonly y: number;
-  };
+  readonly position: Position;
   /**
    * State of the cell
    *