+++ packages/pattern/CHANGELOG.md
@@ -1,3 +1,11 @@
+## [0.2.0] - Unreleased
+
+### Changed
+
+- Add `CACellList.prototype.translate`
+- Add `CACellList.from2dArray`
+- Add `BoundingRect.prototype.union`
+
 ## [0.1.0] - 2025-04-04
 
 ### Changed